refactor(ui): modularize dashboard components
Add reusable UI components and remove unused legacy components.
