+++ admin/ui/src/components/dashboard/authentication-status-overview.tsx
@@ -1,0 +1,9 @@
+export const AuthenticationStatusOverview = () => (
+  <div className="jwt-bg-slate-800 jwt-text-white jwt-rounded-t-xl jwt-p-8 jwt-text-center">
+    <h1 className="jwt-text-2xl jwt-font-bold jwt-text-white">JWT Authentication for WP-API</h1>
+    <p className="jwt-text-slate-300 jwt-mt-4 jwt-max-w-lg jwt-mx-auto jwt-text-base">
+      Enabling stateless authentication for your WordPress REST API with industry-standard JWT
+      tokens.
+    </p>
+  </div>
+)
