+++ admin/ui/src/components/ui/check-icon.tsx
@@ -1,0 +1,3 @@
+import { CheckCircle } from 'lucide-react'
+
+export const Check = () => <CheckCircle className="jwt-h-5 jwt-w-5 jwt-text-emerald-500" />
